# docs/swagger.yaml
swagger: '2.0'
info:
  title: Predico - Restful-API
  description: |

    #### Description:
    Restful API for the INESC TEC Data Sharing / Barter Incentives Service.

    This API is crafted to enable secure data exchange and incentivized collaboration among multiple agents and a central server, where the execution of Data Sharing/Barter Incentives algorithms takes place. It robustly supports internal mechanisms for Data Sharing/Barter Incentives, addressing a diverse range of use cases.

    A prominent use case, characterized by a higher Technology Readiness Level (TRL), is the collaborative forecasting scenario.
    In this context, various agents can trade historical (or future) time-series data—such as power generation or weather measurements data—and acquire time-series forecasts improved by the collective pool of data available on the platform. Notably, agents can access these forecasts without directly tapping into internal datasets. The economic value of data shared by one agent is inherently tied to its utility for other participants that also buy collaborative forecasts from this service.

    ## Main functions:
    - **User management**:
        * Register new user
        * User verification / authentication
        * User password reset
        * Manage User Resources (i.e., bidding units in user portfolio) (create/update/delete)
        * Manage User IOTA wallet address reference (create/update/delete)

    - **Market management**:
...

...

...
  license:
...
  version: v1
host: predico02.inesctec.pt
schemes:
...
basePath: /api
consumes:
...
produces:
...
securityDefinitions:
  Bearer:
    type: apiKey
    name: Authorization
    in: header
security:
  - Bearer: []
paths:
...
  /user/list/:
    get:
      operationId: get_user_list
      description: '[AdminOnly] List current agents details'
      parameters: []
      responses:
        '200':
          description: ''
        '400':
          description: Bad request
        '401':
          description: Unauthorized
          schema:
            title: Not Authenticated schema
            type: object
            properties:
              code:
                description: Response status code.
                type: integer
                enum:
                  - 401
              data:
                description: Null field in case of error response.
                type: object
              status:
                description: Response status info.
                type: string
                enum:
                  - error
              message:
                description: Error message.
                type: string
                enum:
                  - Authentication credentials were not provided.
          examples:
            application/json:
              code: 401
              data: null
              status: error
              message: Given token not valid for any token type
        '403':
          description: Forbidden
          schema:
            title: Forbidden access schema
            type: object
            properties:
              code:
                description: Response status code.
                type: integer
                enum:
                  - 403
              data:
                description: Null field in case of error response.
                type: object
              status:
                description: Response status info.
                type: string
                enum:
                  - error
              message:
                description: Error message.
                type: string
                enum:
                  - You do not have permission to perform this action.
          examples:
            application/json:
              code: 403
              data: null
              status: error
              message: You do not have permission to perform this action.
        '500':
          description: Internal Server Error
      tags:
        - user
    parameters: []
...
  /user/register/:
    post:
      operationId: post_user_register
      description: '[Public] Method for new agent registration. An email is issued with validation link upon registration.'
      parameters:
        - name: data
          in: body
          required: true
          schema:
            $ref: '#/definitions/UserRegistration'
      responses:
        '201':
          description: ''
          schema:
            $ref: '#/definitions/UserRegistration'
        '400':
          description: Bad request
        '500':
          description: Internal Server Error
      tags:
        - user
    parameters: []
...
definitions:
...
  UserRegistration:
    required:
      - email
      - password
      - first_name
      - last_name
    type: object
    properties:
      email:
        title: Email
        type: string
        minLength: 1
      password:
        title: password
        type: string
        minLength: 1
      first_name:
        title: First name
        type: string
        maxLength: 30
        minLength: 1
      last_name:
        title: Last name
        type: string
        maxLength: 30
        minLength: 1
...